:recycle: Refactor the extra data logged
The log item already has a reference to the list item, so including
the status adds little value. However, a human-readable representation
of which zaak failed/succeeded to delete is hard to come by - the URL
alone says nothing, so that should be logged.

We have to be careful to not include any case details themselves,
otherwise we just duplicated the content instead of destroying it.